backend/requirements.txt

fastapi
uvicorn[standard]
httpx
psycopg2-binary
python-dotenv
pydantic
SQLAlchemy==1.4.49
databases
